+++ docker/pip-install.sh
@@ -2,14 +2,13 @@
 
 # Create Python virtual environment if not present
 if [ ! -d /opt/openoni/ENV/lib ]; then
-  pip3 install -U pip
-  pip install -U setuptools
-  pip install -U virtualenv
-  virtualenv -p python3 ENV
+  python3 -m venv ENV
 fi
 
 # Activate the Python virtual environment
 source ENV/bin/activate
 
 # Install Open ONI dependencies
-pip install -r requirements.lock
+# --no-cache-dir disables building local wheels as packages are installed
+# Building wheels provides no benefit when isolated within virtual environment
+pip install --no-cache-dir -r requirements.lock